+++ apps/dashboard/src/app/(main)/integrations/neon/projects/transfer/confirm/page-client.tsx
@@ -3,10 +3,11 @@
 import { Logo } from "@/components/logo";
 import { useRouter } from "@/components/router";
 import { useStackApp, useUser } from "@stackframe/stack";
-import { wait } from "@stackframe/stack-shared/dist/utils/promises";
+import { runAsynchronously, wait } from "@stackframe/stack-shared/dist/utils/promises";
 import { Button, Card, CardContent, CardFooter, CardHeader, Input, Typography } from "@stackframe/stack-ui";
 import Image from "next/image";
 import { useSearchParams } from "next/navigation";
+import { useEffect, useState } from "react";
 import NeonLogo from "../../../../../../../../public/neon.png";
 
 export const stackAppInternalsSymbol = Symbol.for("StackAuth--DO-NOT-USE-OR-YOU-WILL-BE-FIRED--StackAppInternals");
@@ -16,6 +17,28 @@ export default function NeonIntegrationProjectTransferConfirmPageClient() {
   const user = useUser({ projectIdMustMatch: "internal" });
   const router = useRouter();
   const searchParams = useSearchParams();
+
+  const [state, setState] = useState<'loading'|'success'|{type: 'error', message: string}>('loading');
+
+  useEffect(() => {
+    runAsynchronously(async () => {
+      try {
+        await (app as any)[stackAppInternalsSymbol].sendRequest("/integrations/neon/projects/transfer/confirm/check", {
+          method: "POST",
+          body: JSON.stringify({
+            code: searchParams.get("code"),
+          }),
+          headers: {
+            "Content-Type": "application/json",
+          },
+        });
+        setState('success');
+      } catch (err: any) {
+        setState({ type: 'error', message: err.message });
+      }
+    });
+
+  }, [app, searchParams]);
 
   const currentUrl = new URL(window.location.href);
   const signUpSearchParams = new URLSearchParams();
@@ -64,27 +87,35 @@ export default function NeonIntegrationProjectTransferConfirmPageClient() {
         <h1 className="text-3xl font-semibold">
           Project transfer
         </h1>
-        <Typography className="text-sm">
-          Neon would like to transfer a Stack Auth project and link it to your own account. This will let you access the project from Stack Auth&apos;s dashboard.
-        </Typography>
-        {user ? (
-          <>
-            <Typography className="mb-3 text-sm">
-              Which Stack Auth account would you like to transfer the project to? (This is <span className="font-semibold text-red-500">irreversible</span>.)
-            </Typography>
-            <Input type="text" disabled prefixItem={<Logo noLink width={15} height={15} />} value={`Signed in as ${user.primaryEmail || user.displayName || "Unnamed user"}`} />
-            <Button variant="secondary" onClick={async () => await user.signOut({ redirectUrl: signUpUrl })}>
-              Switch account
-            </Button>
-          </>
-        ) : (
+        {state === 'success' && <>
           <Typography className="text-sm">
-            To continue, please sign in or create a Stack Auth account.
+          Neon would like to transfer a Stack Auth project and link it to your own account. This will let you access the project from Stack Auth&apos;s dashboard.
           </Typography>
-        )}
+          {user ? (
+            <>
+              <Typography className="mb-3 text-sm">
+              Which Stack Auth account would you like to transfer the project to? (This is <span className="font-semibold text-red-500">irreversible</span>.)
+              </Typography>
+              <Input type="text" disabled prefixItem={<Logo noLink width={15} height={15} />} value={`Signed in as ${user.primaryEmail || user.displayName || "Unnamed user"}`} />
+              <Button variant="secondary" onClick={async () => await user.signOut({ redirectUrl: signUpUrl })}>
+              Switch account
+              </Button>
+            </>
+          ) : (
+            <Typography className="text-sm">
+            To continue, please sign in or create a Stack Auth account.
+            </Typography>
+          )}
+        </>}
+
+        {typeof state !== 'string' && <>
+          <Typography className="text-sm">
+            {state.message}
+          </Typography>
+        </>}
 
       </CardContent>
-      <CardFooter className="flex justify-end mt-4">
+      {state === 'success' && <CardFooter className="flex justify-end mt-4">
         <div className="flex gap-2 justify-center">
           <Button variant="secondary" onClick={() => { window.close(); }}>
             Cancel
@@ -111,7 +142,7 @@ export default function NeonIntegrationProjectTransferConfirmPageClient() {
             {user ? "Transfer" : "Sign in"}
           </Button>
         </div>
-      </CardFooter>
+      </CardFooter>}
     </Card>
   );
 }