+++ packages/stack-emails/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/stack-emails
 
+## 2.7.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.14
+  - @stackframe/stack-ui@2.7.14
+
 ## 2.7.13
 
 ### Patch Changes