+++ packages/template/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @stackframe/stack
 
+## 2.7.14
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.14
+  - @stackframe/stack-ui@2.7.14
+  - @stackframe/stack-sc@2.7.14
+
 ## 2.7.13
 
 ### Patch Changes